lexer flakey test, moved some code around, did some clippy stuff
