+++ steel/Cargo.toml
@@ -19,7 +19,7 @@ logos = "0.11.4"
 itertools = "0.9"
 log = "0.4"
 futures = "0.3"
-reqwest = { version = "0.10", features = ["json"] }
+# reqwest = { version = "0.10", features = ["json"] }
 async-compat = "0.1.3"
 serde_json = "1.0"
